Add: scripts_all_pairs - new functionality
Allows to run node-to-node tests for all
possible pairs of nodes, uses as many nodes
at once as possible, while not allowing any
node to do two operations at the same time.

Example:

nodes:
  1, 2, 3, 4
scripts:
  iperf-start.sh, iperf-client.sh, iperf-stop.sh

rq structure:

scripts_all_pairs:
  __default:
    - server_start: iperf-start.sh
      server_stop: iperf-stop.sh
      client: iperf-client.sh

the new function run_scripts_all_pairs will:
1. run "iperf-start.sh" on all nodes and record
outputfile from stdout (required to leave process
in background and disconnect). This script
must allocate a temporary file (mktemp),
redirect the output of the desired server
process to this file and use "&" to send it to
background, and it has to print the temporary
file name to stdout (and nothing else,
should be one line only)
2. run "iperf-client.sh" on all nodes in sets,
like so:
  - on 1, providing "SERVER_IP" env var with IP of 2
    on 3, providing "SERVER_IP" with IP of 4
  - on 1, pointing to 3
    on 2, pointing to 4
  - on 1, pointing to 4
    on 2, pointing to 3
3. run "iperf-stop.sh" on all nodes, collecting
(cat-ing) the outputfile recored on step 1

Main purpose of this is to test network
We can do ping or anything else, we can also
skip server part by not defining it

Currenty it uses admin net IPs, need to improve
to allow flexibility (mgmt, storage, any bridge)

Change-Id: Ic3df7cd607aad129c8eb8131f408ee6f8638e084