fix: tummy stuck when calculating log size
Timmy can get stuck on proc.join() if queue has not been collected
and the queue does not fit system pipe buffer between the processes.
This fix adds a loop in which we try to get data from the queue and
then join the process but with a 1 second timeout, so we don't block
forever.

Closes-Bug: https://github.com/adobdin/timmy/issues/66
Change-Id: I0d96a7c8aa6823bc5c22c4b6b74e4897d52d4332