Change: fuel postgres dump collection
We now attempt to collect the dump disregarding the release
For both postgres in docker and in main system.
Two files will always be present in collected results:
fuel-docker-postgres-backup and fuel-postgres-backup
One of them will be empty, the other will contain the backup.

Which one is empty depends on whether postgres is in docker or not.

Change-Id: I9627fb00595ae8ce21188727b8d3de71b64fbcf1